adding new feature paiement

+++ src/main/java/org/example/easymart/dto/request/PromoCodeDTO.java
@@ -3,9 +3,11 @@ package org.example.easymart.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PromoCodeDTO {
     private Long id;
     private String code;
-    private Long pourcentage;
+    private BigDecimal pourcentage;
 }
