+++ src/main/java/org/example/easymart/dto/response/PromoCodeDtoResponse.java
@@ -2,9 +2,11 @@
 
     import lombok.Data;
 
+    import java.math.BigDecimal;
+
     @Data
     public class PromoCodeDtoResponse {
         private Long id;
         private String code;
-        private Long pourcentage;
+        private BigDecimal pourcentage;
     }
